+++ libs/tokens/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [3.8.5](https://github.com/cowprotocol/cowswap/compare/tokens-v3.8.4...tokens-v3.8.5) (2026-07-22)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.9.0
+  - @cowprotocol/ui@3.9.3
+  - @cowprotocol/common-hooks@3.6.1
+  - @cowprotocol/core@3.6.2
+
 ## [3.8.4](https://github.com/cowprotocol/cowswap/compare/tokens-v3.8.3...tokens-v3.8.4) (2026-07-17)
 
 ### Patch Changes